+++ src/plugin/specToolkitPlugin.ts
@@ -1,3 +1,5 @@
+import { PluginOptions } from "../index.js";
+
 export default abstract class SpecToolkitPlugin {
   private readonly _name: string;
 
@@ -9,6 +11,6 @@ export default abstract class SpecToolkitPlugin {
     return this._name;
   }
 
-  public abstract generate(mainSpecSourceFilePaths: string[], outputPath: string, pluginOptions?: unknown): void;
+  public abstract generate(mainSpecSourceFilePaths: string[], outputPath: string, pluginOptions?: PluginOptions): void;
   public abstract get xProperties(): string[] | undefined;
 }